+++ app/src/AREALITY/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.areality">
+    package="areality.game">
 
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
@@ -22,19 +22,19 @@
             android:name="com.google.android.gms.version"
             android:value="@integer/google_play_services_version" />
 
-        <activity android:name=".MainActivity" />
-        <activity android:name=".LoadingScreenActivity">
+        <activity android:name="areality.game.MainActivity" />
+        <activity android:name="areality.game.LoadingScreenActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MapsActivity"></activity>
-        <activity android:name=".LandmarkPage"></activity>
-        <activity android:name=".SignupActivity" />
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".ProfileActivity"></activity>
+        <activity android:name="areality.game.MapsActivity"></activity>
+        <activity android:name="areality.game.LandmarkPage"></activity>
+        <activity android:name="areality.game.SignupActivity" />
+        <activity android:name="areality.game.LoginActivity" />
+        <activity android:name="areality.game.ProfileActivity"></activity>
     </application>
 
 </manifest>
