+++ app/src/AREALITY/app/src/main/java/areality/game/GetNearbyPlacesData.java
@@ -1,0 +1,89 @@
+package areality.game;
+
+import android.os.AsyncTask;
+import android.util.Log;
+
+import com.google.android.gms.common.api.GoogleApiClient;
+import com.google.android.gms.maps.CameraUpdateFactory;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.model.BitmapDescriptorFactory;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.MarkerOptions;
+
+import java.util.HashMap;
+import java.util.List;
+
+/**
+ * Created by tyler on 4/5/17.
+ */
+
+public class GetNearbyPlacesData extends AsyncTask<Object, String, String> {
+    private String googlePlacesData;
+    private GoogleMap mMap;
+    private GoogleApiClient mGoogleApiClient;
+
+    private List<HashMap<String, String>> nearbyPlacesList;
+
+
+    public GetNearbyPlacesData(GoogleApiClient googleApiClient){
+        mGoogleApiClient = googleApiClient;
+    }
+
+    public List<HashMap<String, String>> getPlacesList() {
+        return nearbyPlacesList;
+    }
+
+    @Override
+    protected String doInBackground(Object... params) {
+        try {
+            Log.d("GetNearbyPlacesData", "doInBackground entered");
+            mMap = (GoogleMap) params[0];
+            String url;
+            url = (String) params[1];
+            DownloadUrl downloadUrl = new DownloadUrl();
+            googlePlacesData = downloadUrl.readUrl(url);
+            Log.d("GooglePlacesReadTask", "doInBackground Exit");
+        } catch (Exception e) {
+            Log.d("GooglePlacesReadTask", e.toString());
+        }
+        return googlePlacesData;
+    }
+
+    @Override
+    protected void onPostExecute(String result) {
+        Log.d("GooglePlacesReadTask", "onPostExecute Entered");
+//        List<HashMap<String, String>> nearbyPlacesList = null;
+        DataParser dataParser = new DataParser();
+        nearbyPlacesList = dataParser.parse(result);
+        ShowNearbyPlaces(nearbyPlacesList);
+        Log.d("GooglePlacesReadTask", "onPostExecute Exit");
+    }
+
+    private void ShowNearbyPlaces(List<HashMap<String, String>> nearbyPlacesList) {
+        for (int i = 0; i < nearbyPlacesList.size(); i++) {
+            Log.d("onPostExecute", "Entered into showing locations");
+            MarkerOptions markerOptions = new MarkerOptions();
+            HashMap<String, String> googlePlace = nearbyPlacesList.get(i);
+            double lat = Double.parseDouble(googlePlace.get("lat"));
+            double lng = Double.parseDouble(googlePlace.get("lng"));
+            String placeName = googlePlace.get("place_name");
+            final String placeId = googlePlace.get("place_id");
+            LatLng latLng = new LatLng(lat, lng);
+            markerOptions.position(latLng);
+            markerOptions.title(placeName + ": " + placeId);
+            mMap.addMarker(markerOptions);
+            markerOptions.icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_RED));
+            //move map camera
+
+            mMap.moveCamera(CameraUpdateFactory.newLatLng(latLng));
+            mMap.animateCamera(CameraUpdateFactory.zoomTo(11));
+            
+
+        }
+
+    }
+
+
+
+}
+
