+++ app/src/AREALITY/app/src/main/java/areality/game/MainActivity.java
@@ -1,0 +1,12 @@
+package areality.game;
+
+import android.app.Activity;
+import android.os.Bundle;
+
+public class MainActivity extends Activity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+}
