--- app/src/AREALITY/app/src/main/java/com/example/areality/DataParser.java
@@ -1,87 +1,0 @@
-package com.example.areality;
-
-
-import android.util.Log;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-
-/**
- * Created by tyler on 4/5/17.
- */
-
-public class DataParser {
-    public List<HashMap<String, String>> parse(String jsonData) {
-        JSONArray jsonArray = null;
-        JSONObject jsonObject;
-
-        try {
-            Log.d("Places", "parse");
-            jsonObject = new JSONObject((String) jsonData);
-            jsonArray = jsonObject.getJSONArray("results");
-        } catch (JSONException e) {
-            Log.d("Places", "parse error");
-            e.printStackTrace();
-        }
-        return getPlaces(jsonArray);
-    }
-
-    private List<HashMap<String, String>> getPlaces(JSONArray jsonArray) {
-        int placesCount = jsonArray.length();
-        List<HashMap<String, String>> placesList = new ArrayList<>();
-        HashMap<String, String> placeMap = null;
-        Log.d("Places", "getPlaces");
-
-        for (int i = 0; i < placesCount; i++) {
-            try {
-                placeMap = getPlace((JSONObject) jsonArray.get(i));
-                placesList.add(placeMap);
-                Log.d("Places", "Adding places");
-
-            } catch (JSONException e) {
-                Log.d("Places", "Error in Adding places");
-                e.printStackTrace();
-            }
-        }
-        return placesList;
-    }
-
-    private HashMap<String, String> getPlace(JSONObject googlePlaceJson) {
-        HashMap<String, String> googlePlaceMap = new HashMap<String, String>();
-        String placeName = "-NA-";
-        String latitude = "";
-        String longitude = "";
-        String reference = "";
-        String placeId = "";
-        Log.d("getPlace", "Entered");
-
-        try {
-            if (!googlePlaceJson.isNull("name")) {
-                placeName = googlePlaceJson.getString("name");
-            }
-
-            if (!googlePlaceJson.isNull("place_id")) {
-                placeId = googlePlaceJson.getString("place_id");
-            }
-            latitude = googlePlaceJson.getJSONObject("geometry").getJSONObject("location").getString("lat");
-            longitude = googlePlaceJson.getJSONObject("geometry").getJSONObject("location").getString("lng");
-            reference = googlePlaceJson.getString("reference");
-            googlePlaceMap.put("place_name", placeName);
-            googlePlaceMap.put("place_id", placeId);
-            googlePlaceMap.put("lat", latitude);
-            googlePlaceMap.put("lng", longitude);
-            googlePlaceMap.put("reference", reference);
-            Log.d("getPlace", "Putting Places");
-        } catch (JSONException e) {
-            Log.d("getPlace", "Error");
-            e.printStackTrace();
-        }
-        return googlePlaceMap;
-    }
-
-}
