--- app/src/AREALITY/app/src/main/java/com/example/areality/MainActivity.java
@@ -1,12 +1,0 @@
-package com.example.areality;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-public class MainActivity extends Activity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-}
