+++ app/src/AREALITY/app/src/main/res/layout/activity_loading_screen.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/loading_screen_image"
-    tools:context="com.example.areality.LoadingScreenActivity">
+    tools:context="areality.game.LoadingScreenActivity">
 
 
 </android.support.constraint.ConstraintLayout>
