+++ app/src/AREALITY/app/src/main/res/layout/activity_maps.xml
@@ -32,7 +32,7 @@
             android:name="com.google.android.gms.maps.SupportMapFragment"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context="com.example.areality.MapsActivity"
+            tools:context="com.example.areality.areality.MapsActivity"
             tools:layout="@layout/activity_maps" />
 
         <View
