+++ app/src/AREALITY/app/src/main/res/layout/activity_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.areality.ProfileActivity">
+    tools:context="areality.game.ProfileActivity">
 
     <ScrollView
         android:layout_width="fill_parent"
